requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
gspread
google-auth